+++ src/Milestones/resources/js/data/attributes.js
@@ -28,6 +28,10 @@ const blockAttributes = {
 		type: 'string',
 		default: '',
 	},
+	goal: {
+		type: 'string',
+		default: '',
+	},
 };
 
 export default blockAttributes;