+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Milestone can now be associated with one or many forms (#5230)
 -   Milestone now displays aggregated earnings based on associated forms (#5236)
 -   Milestone now supports a deadline (#5239)
+-   Milestone now supports a custom goal (#5237)
 
 ## [2.8.0] - 2020-08-31